+++ packages/builder-react-core/index.ts
@@ -1,5 +1,3 @@
 export * from "./export-result";
-export * from "./react-component-exporting";
 export * from "./react-import-specifications";
-export * from "./react-module-file";
 export * from "./react-module";
