+++ packages/builder-react-native/rn-build-inline-style-widget/rn-inline-style-module-builder.ts
@@ -7,7 +7,6 @@ import type { JsxWidget } from "@web-builder/core";
 import {
   react_imports,
   ReactWidgetModuleExportable,
-  makeReactModuleFile,
 } from "@web-builder/react-core";
 import {
   buildJsx,
@@ -30,6 +29,7 @@ import {
 import { cssToJson } from "@web-builder/styles/_utils";
 import { CSSProperties } from "@coli.codes/css";
 import { reactnative_imports } from "..";
+import { makeEsWidgetModuleFile } from "@web-builder/module-es";
 
 /**
  * CSS In JS Style builder for React Framework
@@ -182,7 +182,7 @@ export class ReactNativeInlineStyleWidgetModuleExportable extends ReactWidgetMod
   }: {
     exporting: react_config.ReactComponentExportingCofnig;
   }) {
-    return makeReactModuleFile({
+    return makeEsWidgetModuleFile({
       name: this.name,
       path: "src/components",
       imports: this.imports,
