+++ packages/builder-react-native/rn-build-styled-component-widget/rn-styled-components-module-builder.ts
@@ -3,7 +3,6 @@ import { ReservedKeywordPlatformPresets } from "@coli.codes/naming/reserved";
 import { BlockStatement, ImportDeclaration, Return } from "coli";
 import {
   react_imports,
-  makeReactModuleFile,
   ReactWidgetModuleExportable,
   emotion_styled_imports,
   styled_components_imports,
@@ -25,6 +24,7 @@ import {
   StyledComponentDeclaration,
   create_duplication_reduction_map,
 } from "@web-builder/styled";
+import { makeEsWidgetModuleFile } from "@web-builder/module-es";
 
 export class ReactNativeStyledComponentsModuleBuilder {
   private readonly entry: JsxWidget;
@@ -170,7 +170,7 @@ export class ReactNativeStyledComponentWidgetModuleExportable extends ReactWidge
   }: {
     exporting: react_config.ReactComponentExportingCofnig;
   }) {
-    return makeReactModuleFile({
+    return makeEsWidgetModuleFile({
       name: this.name,
       path: "src/components",
       imports: this.imports,
