+++ packages/builder-web-react/react-css-module-widget/react-css-module-module-builder.ts
@@ -12,7 +12,6 @@ import {
 } from "coli";
 import {
   react_imports,
-  makeReactModuleFile,
   ReactWidgetModuleExportable,
 } from "@web-builder/react-core";
 import { JsxWidget } from "@web-builder/core";
@@ -25,6 +24,7 @@ import {
 } from "@web-builder/core/builders";
 import { react as react_config } from "@designto/config";
 import { create_duplication_reduction_map } from "@web-builder/styled";
+import { makeEsWidgetModuleFile } from "@web-builder/module-es";
 
 /**
  * CSS Module Builder for React Framework
@@ -182,7 +182,7 @@ export class ReactCssModuleWidgetModuleExportable extends ReactWidgetModuleExpor
   }: {
     exporting: react_config.ReactComponentExportingCofnig;
   }) {
-    return makeReactModuleFile({
+    return makeEsWidgetModuleFile({
       name: this.name,
       path: "src/components",
       imports: this.imports,
