+++ packages/builder-web-react/react-inline-css-widget/react-inline-css-module-builder.ts
@@ -4,7 +4,6 @@ import type { JsxWidget } from "@web-builder/core";
 import {
   react_imports,
   ReactWidgetModuleExportable,
-  makeReactModuleFile,
 } from "@web-builder/react-core";
 import {
   buildJsx,
@@ -26,6 +25,7 @@ import {
 } from "coli";
 import { cssToJson } from "@web-builder/styles/_utils";
 import { CSSProperties } from "@coli.codes/css";
+import { makeEsWidgetModuleFile } from "@web-builder/module-es";
 
 /**
  * InlineCss Style builder for React Framework
@@ -174,7 +174,7 @@ export class ReactInlineCssWidgetModuleExportable extends ReactWidgetModuleExpor
   }: {
     exporting: react_config.ReactComponentExportingCofnig;
   }) {
-    return makeReactModuleFile({
+    return makeEsWidgetModuleFile({
       name: this.name,
       path: "src/components",
       imports: this.imports,
