+++ packages/builder-web-react/react-styled-component-widget/react-styled-components-module-builder.ts
@@ -8,7 +8,6 @@ import {
 } from "@web-builder/styled";
 import {
   react_imports,
-  makeReactModuleFile,
   ReactWidgetModuleExportable,
   emotion_styled_imports,
   styled_components_imports,
@@ -25,6 +24,7 @@ import {
   StyledComponentDeclaration,
   create_duplication_reduction_map,
 } from "@web-builder/styled";
+import { makeEsWidgetModuleFile } from "@web-builder/module-es";
 
 export class ReactStyledComponentsBuilder {
   private readonly entry: JsxWidget;
@@ -162,7 +162,7 @@ export class ReactStyledComponentWidgetModuleExportable extends ReactWidgetModul
   }: {
     exporting: react_config.ReactComponentExportingCofnig;
   }) {
-    return makeReactModuleFile({
+    return makeEsWidgetModuleFile({
       name: this.name,
       path: "src/components",
       imports: this.imports,
